+++ src/axom/slam/policies/IndirectionPolicies.hpp
@@ -33,6 +33,7 @@
 #define SLAM_POLICIES_INDIRECTION_H_
 
 #include "axom/core/Macros.hpp"
+#include "axom/core/Array.hpp"
 #include "axom/slic/interface/slic.hpp"
 
 namespace axom
@@ -58,6 +59,7 @@ struct NoIndirection
   using ConstIndirectionResult = const ElementType;
   using IndirectionBufferType = struct
   { };
+  using IndirectionPtrType = IndirectionBufferType*;
 
   NoIndirection() { }
 
@@ -93,6 +95,7 @@ struct ArrayIndirection
   using ConstIndirectionResult = const ElementType&;
 
   using IndirectionBufferType = ElementType;
+  using IndirectionPtrType = IndirectionBufferType*;
 
   ArrayIndirection(IndirectionBufferType* buf = nullptr) : m_arrBuf(buf) { }
 
@@ -146,6 +149,7 @@ struct STLVectorIndirection
 
   using VectorType = std::vector<ElementType>;
   using IndirectionBufferType = VectorType;
+  using IndirectionPtrType = IndirectionBufferType*;
 
   STLVectorIndirection(IndirectionBufferType* buf = nullptr) : m_vecBuf(buf) { }
 
@@ -214,6 +218,7 @@ struct CoreArrayIndirection
 
   using VectorType = axom::Array<ElementType>;
   using IndirectionBufferType = VectorType;
+  using IndirectionPtrType = IndirectionBufferType*;
 
   CoreArrayIndirection(IndirectionBufferType* buf = nullptr) : m_vecBuf(buf) { }
 
@@ -270,6 +275,67 @@ struct CoreArrayIndirection
 
 private:
   IndirectionBufferType* m_vecBuf;
+};
+
+/**
+ * \brief A policy class for sets with axom::ArrayView-based indirection
+ */
+template <typename PositionType, typename ElementType>
+struct ArrayViewIndirection
+{
+  using IndirectionResult = ElementType&;
+  using ConstIndirectionResult = const ElementType&;
+
+  using VectorType = axom::ArrayView<ElementType>;
+  using IndirectionBufferType = VectorType;
+  using IndirectionPtrType = IndirectionBufferType;
+
+  ArrayViewIndirection(IndirectionBufferType buf = {}) : m_vecBuf(buf) { }
+
+  IndirectionBufferType& data() { return m_vecBuf; }
+  const IndirectionBufferType& data() const { return m_vecBuf; }
+
+  inline ConstIndirectionResult indirection(PositionType pos) const
+  {
+    SLIC_ASSERT_MSG(hasIndirection(),
+                    "SLAM::Set:ArrayViewIndirection -- Tried to dereference "
+                      << "a null vector in a vector based indirection set.");
+    //SLIC_ASSERT_MSG( pos < m_vecBuf->size(),
+    //  "SLAM::Set:ArrayViewIndirection -- "
+    //  << "Tried to access an out of bounds element at position "
+    //  << pos << " in vector with only " << m_vecBuf->size() << " elements.");
+
+    return m_vecBuf[pos];
+  }
+
+  inline IndirectionResult indirection(PositionType pos)
+  {
+    SLIC_ASSERT_MSG(hasIndirection(),
+                    "SLAM::Set:ArrayViewIndirection -- Tried to dereference "
+                      << "a null vector in a vector based indirection set.");
+
+    return m_vecBuf[pos];
+  }
+
+  inline IndirectionResult operator()(PositionType pos)
+  {
+    return indirection(pos);
+  }
+
+  inline ConstIndirectionResult operator()(PositionType pos) const
+  {
+    return indirection(pos);
+  }
+
+  bool hasIndirection() const { return m_vecBuf.data() != nullptr; }
+
+  inline bool isValid(PositionType size,
+                      PositionType offset,
+                      PositionType stride,
+                      bool verboseOutput = false) const;
+
+private:
+  IndirectionBufferType m_vecBuf;
 };
 
 /// \}
@@ -408,6 +474,59 @@ bool CoreArrayIndirection<PosType, ElemType>::isValid(PosType size,
     PosType firstEltInd = offset;
     PosType lastEltInd = (size - 1) * stride + offset;
     PosType vecSize = static_cast<PosType>(m_vecBuf->size());
+
+    bool isRangeValid = (0 <= firstEltInd) && (firstEltInd < vecSize) &&
+      (0 <= lastEltInd) && (lastEltInd < vecSize);
+
+    if(!isRangeValid)
+    {
+      SLIC_DEBUG_IF(
+        verboseOutput,
+        "Invalid array-based IndirectionSet -- Data buffer "
+          << "must be large enough to hold all elements of the set. "
+          << "Underlying buffer size is " << vecSize << "."
+          << " Offset of " << offset << " leads to a first index of "
+          << firstEltInd << "."
+          << " Stride of " << stride << " and size of " << size
+          << " leads to a last index of " << lastEltInd << ".");
+
+      bValid = false;
+    }
+  }
+
+  return bValid;
+}
+
+template <typename PosType, typename ElemType>
+bool ArrayViewIndirection<PosType, ElemType>::isValid(PosType size,
+                                                      PosType offset,
+                                                      PosType stride,
+                                                      bool verboseOutput) const
+{
+  AXOM_UNUSED_VAR(verboseOutput);
+
+  // always valid if set has zero size, even if indirection buffer is null
+  if(size == 0) return true;
+
+  bool bValid = true;
+
+  // Otherwise, check whether the set has elements, but the array ptr is null
+  if(!hasIndirection())
+  {
+    SLIC_DEBUG_IF(verboseOutput,
+                  "Array-based indirection set with non-zero size "
+                    << "(size=" << size << ") requires a valid data buffer,"
+                    << "but buffer pointer was null.");
+
+    bValid = false;
+  }
+  else
+  {
+    // Verify underlying vector has sufficient storage for all set elements
+    // Note: it is valid for the data buffer to have extra space
+    PosType firstEltInd = offset;
+    PosType lastEltInd = (size - 1) * stride + offset;
+    PosType vecSize = static_cast<PosType>(m_vecBuf.size());
 
     bool isRangeValid = (0 <= firstEltInd) && (firstEltInd < vecSize) &&
       (0 <= lastEltInd) && (lastEltInd < vecSize);